refactor(runtime): move useNodeViewMessageContext to lib/messages
