+++ ui/src/lib/messages/index.ts
@@ -1,4 +1,7 @@
 export { MessageContext, type SendMessageOptions } from './MessageContext';
 export { MessageSystem, type Message, type MessageCallbackFn } from './MessageSystem';
+
 export { DirectChannelService } from './DirectChannelService';
 export { MessageChannelRegistry } from './MessageChannelRegistry';
+
+export { useNodeViewMessageContext } from './useNodeViewMessageContext.svelte';
